pocketlawdemo/backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
requests==2.32.3
PyPDF2==3.0.1
python-docx==1.1.2
